+++ src/test/java/com/example/ohlccryptoapi/testHelper/TestCryptoCurrencyRepository.java
@@ -1,14 +1,15 @@
 package com.example.ohlccryptoapi.testHelper;
 
-import com.example.ohlccryptoapi.domain.model.security.ohlc.Cryptocurrency;
-import com.example.ohlccryptoapi.domain.repository.CryptocurrencyRepository;
+import com.example.ohlccryptoapi.domain.model.ohlc.Cryptocurrency;
+import com.example.ohlccryptoapi.domain.repository.ohlc.CryptocurrencyRepository;
+import io.swagger.models.auth.In;
 
 import java.util.List;
 import java.util.Optional;
 
 public class TestCryptoCurrencyRepository implements CryptocurrencyRepository {
     @Override
-    public Optional<Cryptocurrency> findById(long id) {
+    public Optional<Cryptocurrency> findById(Integer id) {
         return Optional.empty();
     }
