+++ src/test/java/com/example/ohlccryptoapi/testHelper/TestCryptoCurrencyService.java
@@ -1,7 +1,7 @@
 package com.example.ohlccryptoapi.testHelper;
 
-import com.example.ohlccryptoapi.domain.model.security.ohlc.Cryptocurrency;
-import com.example.ohlccryptoapi.domain.service.CryptoCurrencyService;
+import com.example.ohlccryptoapi.domain.model.ohlc.Cryptocurrency;
+import com.example.ohlccryptoapi.domain.service.ohlc.CryptoCurrencyService;
 
 import java.util.List;
 import java.util.Optional;
@@ -10,7 +10,7 @@ public class TestCryptoCurrencyService implements CryptoCurrencyService {
     private TestCryptoCurrencyRepository  testCryptoCurrencyRepository = new TestCryptoCurrencyRepository();
 
     @Override
-    public Optional<Cryptocurrency> findCryptoCurrencyById(long id) {
+    public Optional<Cryptocurrency> findCryptoCurrencyById(Integer id) {
         return Optional.empty();
     }
